+++ NEWS.md
@@ -8,6 +8,8 @@
 
 - `kable()` operating in LaTeX mode can now optionally typeset numeric columns in math mode for improved rendering of minus signs, infinite values, and scientific notation; in particular, decimal and thousands separator commas are wrapped in braces (`{}`) to preserve spacing. To enable, use `kable(..., format = "latex", numeric.math = TRUE)` or set `options(knitr.table.numeric.math = TRUE)` globally (thanks, @krivit, #1709).
 
+- `kable()` in LaTeX mode gained a new argument `caption.pos` to control whether the table caption is placed at the `'top'` (the default) or `'bottom'` of a table wrapped in a floating `table` environment. The default can be set globally via `options(knitr.table.caption.pos = 'bottom')` (thanks, @njtierney, #1189).
+
 - A code chunk that returns a `shiny.tag` or `shiny.tag.list` object (e.g., from **htmltools**) can now have a figure caption and be cross-referenced, in the same way as **htmlwidgets**. For example, the chunk below produces a captioned figure that can be referenced via `\@ref(fig:mytag)` in **bookdown** (thanks, @cpsievert, #1650):
 
     ````